+++ sphinx_gallery/gen_gallery.py
@@ -15,12 +15,14 @@ import codecs
 import copy
 import re
 import os
+from copy import deepcopy
 
 from . import sphinx_compatibility, glr_path_static, __version__ as _sg_version
 from .gen_rst import generate_dir_rst, SPHX_GLR_SIG
 from .docs_resolv import embed_code_links
 from .downloads import generate_zipfiles
 from .sorting import NumberOfCodeLinesSortKey
+from .binder import copy_binder_reqs, check_binder_conf
 
 try:
     FileNotFoundError
@@ -49,6 +51,7 @@ DEFAULT_GALLERY_CONF = {
     'expected_failing_examples': set(),
     'thumbnail_size': (400, 280),  # Default CSS does 0.4 scaling (160, 112)
     'min_reported_time': 0,
+    'binder': {}
 }
 
 logger = sphinx_compatibility.getLogger('sphinx-gallery')
@@ -190,7 +193,7 @@ def _prepare_sphx_glr_dirs(gallery_conf, srcdir):
         if not os.path.exists(backreferences_dir):
             os.makedirs(backreferences_dir)
 
-    return zip(examples_dirs, gallery_dirs)
+    return list(zip(examples_dirs, gallery_dirs))
 
 
 def generate_gallery_rst(app):
@@ -209,6 +212,11 @@ def generate_gallery_rst(app):
     computation_times = []
     workdirs = _prepare_sphx_glr_dirs(gallery_conf,
                                       app.builder.srcdir)
+
+    # Check for duplicate filenames to make sure linking works as expected
+    examples_dirs = [ex_dir for ex_dir, _ in workdirs]
+    files = collect_gallery_files(examples_dirs)
+    check_duplicate_filenames(files)
 
     for examples_dir, gallery_dir in workdirs:
 
@@ -257,6 +265,12 @@ def generate_gallery_rst(app):
                     logger.info("\t- %s: %.2g sec", fname, time_elapsed)
             else:
                 logger.info("\t- %s: not run", fname)
+
+    # Copy the requirements files for binder
+    binder_conf = check_binder_conf(gallery_conf.get('binder'))
+    if len(binder_conf) > 0:
+        logger.info("copying binder requirements...")
+        copy_binder_reqs(app)
 
 
 def touch_empty_backreferences(app, what, name, obj, options, lines):
@@ -327,6 +341,36 @@ def sumarize_failing_examples(app, exception):
         raise ValueError("Here is a summary of the problems encountered when "
                          "running the examples\n\n" + "\n".join(fail_msgs) +
                          "\n" + "-" * 79)
+
+
+def collect_gallery_files(examples_dirs):
+    """Collect python files from the gallery example directories."""
+    files = []
+    for example_dir in examples_dirs:
+        for root, dirnames, filenames in os.walk(example_dir):
+            for filename in filenames:
+                if filename.endswith('.py'):
+                    files.append(os.path.join(root, filename))
+    return files
+
+
+def check_duplicate_filenames(files):
+    """Check for duplicate filenames across gallery directories."""
+    # Check whether we'll have duplicates
+    used_names = set()
+    dup_names = list()
+
+    for this_file in files:
+        this_fname = os.path.basename(this_file)
+        if this_fname in used_names:
+            dup_names.append(this_file)
+        else:
+            used_names.add(this_fname)
+
+    if len(dup_names) > 0:
+        logger.warning(
+            'Duplicate file name(s) found. Having duplicate file names will '
+            'break some links. List of files: {}'.format(sorted(dup_names),))
 
 
 def get_default_config_value(key):